Update bundled krb5 distribution to version krb5-1.12.4

+++ krb5/src/lib/rpc/xdr.c
@@ -1,31 +1,35 @@
 /* @(#)xdr.c	2.1 88/07/29 4.0 RPCSRC */
 /*
- * Sun RPC is a product of Sun Microsystems, Inc. and is provided for
- * unrestricted use provided that this legend is included on all tape
- * media and as a part of the software program in whole or part.  Users
- * may copy or modify Sun RPC without charge, but are not authorized
- * to license or distribute it to anyone else except as part of a product or
- * program developed by the user.
- * 
- * SUN RPC IS PROVIDED AS IS WITH NO WARRANTIES OF ANY KIND INCLUDING THE
- * WARRANTIES OF DESIGN, MERCHANTIBILITY AND FITNESS FOR A PARTICULAR
- * PURPOSE, OR ARISING FROM A COURSE OF DEALING, USAGE OR TRADE PRACTICE.
- * 
- * Sun RPC is provided with no support and without any obligation on the
- * part of Sun Microsystems, Inc. to assist in its use, correction,
- * modification or enhancement.
- * 
- * SUN MICROSYSTEMS, INC. SHALL HAVE NO LIABILITY WITH RESPECT TO THE
- * INFRINGEMENT OF COPYRIGHTS, TRADE SECRETS OR ANY PATENTS BY SUN RPC
- * OR ANY PART THEREOF.
- * 
- * In no event will Sun Microsystems, Inc. be liable for any lost revenue
- * or profits or other special, indirect and consequential damages, even if
- * Sun has been advised of the possibility of such damages.
- * 
- * Sun Microsystems, Inc.
- * 2550 Garcia Avenue
- * Mountain View, California  94043
+ * Copyright (c) 2010, Oracle America, Inc.
+ *
+ * All rights reserved.
+ *
+ * Redistribution and use in source and binary forms, with or without
+ * modification, are permitted provided that the following conditions are met:
+ *
+ *     * Redistributions of source code must retain the above copyright
+ *       notice, this list of conditions and the following disclaimer.
+ *
+ *     * Redistributions in binary form must reproduce the above copyright
+ *       notice, this list of conditions and the following disclaimer in
+ *       the documentation and/or other materials provided with the
+ *       distribution.
+ *
+ *     * Neither the name of the "Oracle America, Inc." nor the names of
+ *       its contributors may be used to endorse or promote products
+ *       derived from this software without specific prior written permission.
+ *
+ * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS
+ * IS" AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED
+ * TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A
+ * PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT
+ * HOLDER OR CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL,
+ * SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT LIMITED
+ * TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE, DATA, OR
+ * PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY OF
+ * LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING
+ * NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS
+ * SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 #if !defined(lint) && defined(SCCSIDS)
 static char sccsid[] = "@(#)xdr.c 1.35 87/08/12";
@@ -33,8 +37,6 @@ static char sccsid[] = "@(#)xdr.c 1.35 87/08/12";
 
 /*
  * xdr.c, Generic XDR routines implementation.
- *
- * Copyright (C) 1986, Sun Microsystems, Inc.
  *
  * These are the "generic" xdr routines used to serialize and de-serialize
  * most common data items.  See xdr.h for more info on the interface to
@@ -74,7 +76,7 @@ void
 xdr_free(xdrproc_t proc, void *objp)
 {
 	XDR x;
-	
+
 	x.x_op = XDR_FREE;
 	(*proc)(&x, objp);
 }
@@ -145,7 +147,7 @@ xdr_u_int(XDR *xdrs, u_int *up)
 		if (!XDR_GETLONG(xdrs, (long *) &l))
 			return (FALSE);
 
-		if (l > UINT_MAX)
+		if ((uint32_t)l > UINT_MAX)
 			return (FALSE);
 
 		*up = (u_int) l;
@@ -501,7 +503,7 @@ xdr_int32(XDR *xdrs, int32_t *ip)
 	case XDR_ENCODE:
 		VALGRIND_CHECK_DEFINED(*ip);
 		l = *ip;
-		return (xdr_long(xdrs, &l));    
+		return (xdr_long(xdrs, &l));
 
 	case XDR_DECODE:
 		if (!xdr_long(xdrs, &l)) {
@@ -511,7 +513,7 @@ xdr_int32(XDR *xdrs, int32_t *ip)
 		return (TRUE);
 
 	case XDR_FREE:
-		return (TRUE);    
+		return (TRUE);
 	}
 	return (FALSE);
 }
@@ -526,7 +528,7 @@ xdr_u_int32(XDR *xdrs, uint32_t *up)
 	case XDR_ENCODE:
 		VALGRIND_CHECK_DEFINED(*up);
 		ul = *up;
-		return (xdr_u_long(xdrs, &ul));    
+		return (xdr_u_long(xdrs, &ul));
 
 	case XDR_DECODE:
 		if (!xdr_u_long(xdrs, &ul)) {
@@ -536,7 +538,7 @@ xdr_u_int32(XDR *xdrs, uint32_t *up)
 		return (TRUE);
 
 	case XDR_FREE:
-		return (TRUE);    
+		return (TRUE);
 	}
 	return (FALSE);
 }
@@ -661,8 +663,8 @@ xdr_string(XDR *xdrs, char **cpp, u_int maxsize)
 	return (FALSE);
 }
 
-/* 
- * Wrapper for xdr_string that can be called directly from 
+/*
+ * Wrapper for xdr_string that can be called directly from
  * routines like clnt_call
  */
 bool_t
